+++ services/explorer-ui/src/routes/contracts/index.lazy.tsx
@@ -1,5 +1,5 @@
 import { createLazyFileRoute } from "@tanstack/react-router";
-import { Contracts } from "~/pages/contracts";
+import { Contracts } from "~/pages/contract";
 
 export const Route = createLazyFileRoute("/contracts/")({
   component: Contracts,
